requirements.txt
# Core dependencies
numpy==1.26.2

# API
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
httpx==0.24.1
statsbombpy==1.10.0

# Streamlit
streamlit==1.28.2

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
langchain==0.0.339
google-generativeai==0.3.1
pandas==2.1.3
requests==2.31.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1